pyproject.toml
[project]
name = "yallm"
version = "0.1.1"
description = "Add your description here"
requires-python = ">=3.10,<3.13"
dependencies = [
    "asyncer>=0.0.8",
    "asyncify>=0.11.0",
    "celery>=5.5.3",
    "docx2txt>=0.9",
    "fastapi>=0.115.14",
    "fastapi-cli>=0.0.7",
    "rich>=13.7",
    "huggingface-hub>=0.33.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-redis>=0.2.3",
    "orjson>=3.10.18",
    "pydantic-settings>=2.2.1",
    "pymongo>=4.13.2",
    "motor>=3.7.1",
    "pypdf>=5.7.0",
    "redis>=5.0.4",
    "uvicorn[standard]>=0.35.0",
    "aiogram>=3.5.0",
    "structlog>=24.1.0",
    "qdrant-client>=1.9.0",
    "aiohttp>=3.9.5",
    "uvloop>=0.17.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    # sentence-transformers (and torch) removed for Ollama-only runtime
    "nltk>=3.8.1",
    "prometheus-client>=0.20.0",
    "requests>=2.32,<3",
    "beautifulsoup4>=4.12,<5",
    "lxml>=5.2.0",
    "Pillow>=10.0",
    "openpyxl>=3.1.5",
    "mcp>=1.0.0",
]

[dependency-groups]
dev = [
    "celery-types>=0.23.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "celery-types>=0.23.0",
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = [
    "apps*",
    "packages*",
]

[tool.poetry.source]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu/simple"
